Hoist export files into src folder
